+++ packages/react/test/specs/components/Popup/Popup-test.tsx
@@ -8,6 +8,9 @@ import {
   Alignment,
 } from 'src/components/Popup/positioningHelper'
 import Popup, { PopupEvents } from 'src/components/Popup/Popup'
+import { Accessibility } from 'src/lib/accessibility/types'
+import { popupFocusTrapBehavior, popupBehavior, dialogBehavior } from 'src/lib/accessibility/index'
+
 import { domEvent, mountWithProvider } from '../../../utils'
 import * as keyboardKey from 'keyboard-key'
 import { ReactWrapper } from 'enzyme'
@@ -67,12 +70,6 @@ describe('Popup', () => {
     popupTriggerElement.simulate('keydown', { keyCode: keyboardKeyToOpen })
 
     expect(getPopupContent(popup).exists()).toBe(true)
-
-    // when popup open, check that stopPropagation is called when keyboard events are invoked
-    const stopPropagation = jest.fn()
-    const popupContentElement = getPopupContent(popup)
-    popupContentElement.simulate('keyDown', { stopPropagation })
-    expect(stopPropagation).toHaveBeenCalledTimes(1)
 
     // check popup closes on Esc
     popupTriggerElement.simulate('keydown', { keyCode: keyboardKeyToClose })
@@ -245,6 +242,59 @@ describe('Popup', () => {
       expect(wrapper.find(`#${contentId2}`).exists()).toBe(true)
 
       document.body.removeChild(attachTo)
+    })
+  })
+
+  describe('inline', () => {
+    test('renders the content in the document body the inline prop is not provided', () => {
+      mountWithProvider(<Popup trigger={<button />} content="Content" open={true} />)
+      const contentElement = document.body.firstElementChild
+
+      expect(contentElement.classList.contains(Popup.Content.className)).toEqual(true)
+    })
+
+    test('renders the content next to the trigger element if the inline prop is provided', () => {
+      const wrapper = mountWithProvider(
+        <Popup trigger={<button id={triggerId} />} inline content="Content" open={true} />,
+      )
+      const contentElement = wrapper.getDOMNode().nextSibling as HTMLDivElement
+
+      expect(wrapper.find(Popup.Content).exists()).toEqual(true)
+      expect(contentElement.classList.contains(Popup.Content.className)).toEqual(true)
+    })
+  })
+
+  describe('keyboard event propagation', () => {
+    const expectPopupToHandleStopPropagation = (
+      behavior: Accessibility,
+      shouldStopPropagation: boolean,
+    ) => {
+      const popup = mountWithProvider(
+        <Popup
+          trigger={<span id={triggerId}> text to trigger popup </span>}
+          content={<span id={contentId} />}
+          accessibility={behavior}
+        />,
+      )
+
+      // open popup
+      const popupTriggerElement = popup.find(`#${triggerId}`)
+      popupTriggerElement.simulate('keydown', { keyCode: keyboardKey.Enter })
+
+      // when popup open, check that stopPropagation is called when keyboard events are invoked
+      const stopPropagation = jest.fn()
+      const popupContentElement = getPopupContent(popup)
+      popupContentElement.simulate('keyDown', { stopPropagation })
+      expect(stopPropagation).toHaveBeenCalledTimes(shouldStopPropagation ? 1 : 0)
+    }
+    test('stops when focus trap behavior is used', () => {
+      expectPopupToHandleStopPropagation(popupFocusTrapBehavior, true)
+    })
+    test('stops when dialog behavior is used', () => {
+      expectPopupToHandleStopPropagation(dialogBehavior, true)
+    })
+    test('does not stop when default behavior is used', () => {
+      expectPopupToHandleStopPropagation(popupBehavior, false)
     })
   })
 })